svi: added checks in constructors

+++ src/main/java/it/unibo/smartcity/model/api/TariffaAbbonamento.java
@@ -1,13 +1,11 @@
 package it.unibo.smartcity.model.api;
 
-import java.math.BigDecimal;
-
 public interface TariffaAbbonamento {
 
     String getNome();
 
     int getNum_giorni();
 
-    BigDecimal getPrezzo();
+    Double getPrezzo();
 
 }
